Переносит код блока нового комментария из отдельного компонента в модуль класса FilmDetails.
